Upgrade the release process
This is primarily to get this onto npm, secondarily to get the niceness
of `np` into the mix.

We removed the build and the *.node file/dirs from the package (what
ends up in the tarball) because the user will end up clobbering those
when they build locally via the `prepare` script which gets called
during npm installation time. No point in us packing the tarball with
useless data.